edit '*/[StoneyVCV files]' - relocate to a more install-friendly namespaced dir
