+++ src/StoneyVCV/HP1.cpp
@@ -1,5 +1,5 @@
 /***************************************************************************//**
- * @file HP4.cpp
+ * @file HP1.cpp
  * @author Nathan J. Hood <nathanjhood@googlemail.com>
  * @brief
  * @version 0.0.0
@@ -31,58 +31,54 @@
 
 //==============================================================================
 
-#include "HP4.hpp"
+#include "StoneyVCV/HP1.hpp"
 
 //==============================================================================
 
-::StoneyDSP::StoneyVCV::HP4Module::HP4Module()
+::StoneyDSP::StoneyVCV::HP1Module::HP1Module()
 {
     // Configure the number of Params, Outputs, Inputs, and Lights.
     config(
-        ::StoneyDSP::StoneyVCV::HP4Module::PARAMS_LEN,   // numParams
-        ::StoneyDSP::StoneyVCV::HP4Module::INPUTS_LEN,   // numInputs
-        ::StoneyDSP::StoneyVCV::HP4Module::OUTPUTS_LEN,  // numOutputs
-        ::StoneyDSP::StoneyVCV::HP4Module::LIGHTS_LEN    // numLights
+        ::StoneyDSP::StoneyVCV::HP1Module::PARAMS_LEN,   // numParams
+        ::StoneyDSP::StoneyVCV::HP1Module::INPUTS_LEN,   // numInputs
+        ::StoneyDSP::StoneyVCV::HP1Module::OUTPUTS_LEN,  // numOutputs
+        ::StoneyDSP::StoneyVCV::HP1Module::LIGHTS_LEN    // numLights
     );
 }
 
-// ::StoneyDSP::StoneyVCV::HP4Module::~HP4Module()
+// ::StoneyDSP::StoneyVCV::HP1Module::~HP1Module()
 // {}
 
 //==============================================================================
 
-::StoneyDSP::StoneyVCV::HP4Widget::HP4Widget()
-// :   hp4WidgetFrameBuffer(new ::rack::FramebufferWidget),
+::StoneyDSP::StoneyVCV::HP1Widget::HP1Widget()
+// :   hp1WidgetFrameBuffer(new ::rack::FramebufferWidget),
 //     panelBorder(::rack::createWidget<::rack::PanelBorder>(::rack::math::Vec(0.0f, 0.0f)))
 {
     // // Widgets
-    // hp4WidgetFrameBuffer = new ::rack::FramebufferWidget;
-    // hp4WidgetFrameBuffer->setSize(box.size);
-    // addChild(hp4WidgetFrameBuffer);
+    // hp1WidgetFrameBuffer = new ::rack::FramebufferWidget;
+    // hp1WidgetFrameBuffer->setSize(box.size);
+    // addChild(hp1WidgetFrameBuffer);
     // //
     // panelBorder = ::rack::createWidget<::rack::PanelBorder>(::rack::math::Vec(0.0f, 0.0f));
     // panelBorder->setSize(box.size);
-    // hp4WidgetFrameBuffer->addChild(panelBorder);
+    // hp1WidgetFrameBuffer->addChild(panelBorder);
 }
 
-// ::StoneyDSP::StoneyVCV::HP4Widget::~HP4Widget()
+// ::StoneyDSP::StoneyVCV::HP1Widget::~HP1Widget()
 // {}
 
-void ::StoneyDSP::StoneyVCV::HP4Widget::step()
+void ::StoneyDSP::StoneyVCV::HP1Widget::step()
 {
     // panelBorder->box.size = box.size;
     ::rack::Widget::step();
 }
 
-void ::StoneyDSP::StoneyVCV::HP4Widget::draw(const ::rack::Widget::DrawArgs &args)
+void ::StoneyDSP::StoneyVCV::HP1Widget::draw(const ::rack::Widget::DrawArgs &args)
 {
-    NVGcolor bgBlack = ::StoneyDSP::StoneyVCV::Panels::bgBlack;
-    NVGcolor bgWhite = ::StoneyDSP::StoneyVCV::Panels::bgWhite;
-
-    //
     ::nvgBeginPath(args.vg);
     ::nvgRect(args.vg, 0.0, 0.0, box.size.x, box.size.y);
-    ::NVGcolor bg = ::rack::settings::preferDarkPanels ? bgBlack : bgWhite;
+    ::NVGcolor bg = ::rack::settings::preferDarkPanels ? ::nvgRGB(42, 42, 42) : ::nvgRGB(235, 235, 235);
     ::nvgFillColor(args.vg, bg);
     ::nvgFill(args.vg);
     ::rack::Widget::draw(args);
@@ -90,29 +86,29 @@ void ::StoneyDSP::StoneyVCV::HP4Widget::draw(const ::rack::Widget::DrawArgs &arg
 
 //==============================================================================
 
-::StoneyDSP::StoneyVCV::HP4ModuleWidget::HP4ModuleWidget(::StoneyDSP::StoneyVCV::HP4Module* module)
-// :   hp4Widget(::rack::createWidget<::StoneyDSP::StoneyVCV::HP4Widget>(::rack::math::Vec(0.0F, 0.0F))),
-//     hp4ModuleWidgetFrameBuffer(new ::rack::FramebufferWidget)
+::StoneyDSP::StoneyVCV::HP1ModuleWidget::HP1ModuleWidget(::StoneyDSP::StoneyVCV::HP1Module* module)
+// :   hp1Widget(::rack::createWidget<::StoneyDSP::StoneyVCV::HP1Widget>(::rack::math::Vec(0.0F, 0.0F))),
+//     hp1ModuleWidgetFrameBuffer(new ::rack::FramebufferWidget)
 {
     setModule(module);
     setPanel(::rack::createPanel(
         // Light-mode panel
         ::rack::asset::plugin(
-            ::StoneyDSP::StoneyVCV::pluginInstance, "res/HP4-light.svg"
+            ::StoneyDSP::StoneyVCV::pluginInstance, "res/HP1-light.svg"
         ),
         // Dark-mode panel
         ::rack::asset::plugin(
-            ::StoneyDSP::StoneyVCV::pluginInstance, "res/HP4-dark.svg"
+            ::StoneyDSP::StoneyVCV::pluginInstance, "res/HP1-dark.svg"
         )
     ));
     // // Widgets
-    // hp4ModuleWidgetFrameBuffer = new ::rack::FramebufferWidget;
-    // hp4ModuleWidgetFrameBuffer->setSize(box.size);
-    // addChild(hp4ModuleWidgetFrameBuffer);
+    // hp1ModuleWidgetFrameBuffer = new ::rack::FramebufferWidget;
+    // hp1ModuleWidgetFrameBuffer->setSize(box.size);
+    // addChild(hp1ModuleWidgetFrameBuffer);
     // //
-    // hp4Widget = ::rack::createWidget<::StoneyDSP::StoneyVCV::HP4Widget>(::rack::math::Vec(0.0F, 0.0F));
-    // hp4Widget->setSize(box.size);
-    // hp4ModuleWidgetFrameBuffer->addChild(hp4Widget);
+    // hp1Widget = ::rack::createWidget<::StoneyDSP::StoneyVCV::HP1Widget>(::rack::math::Vec(0.0F, 0.0F));
+    // hp1Widget->setSize(box.size);
+    // hp1ModuleWidgetFrameBuffer->addChild(hp1Widget);
     // // Screws
     // ::rack::math::Vec screwT1Pos = ::rack::math::Vec(::rack::RACK_GRID_WIDTH, 0.0F); // top-middle
     // ::rack::math::Vec screwB1Pos = ::rack::math::Vec(::rack::RACK_GRID_WIDTH, ::rack::RACK_GRID_HEIGHT - ::rack::RACK_GRID_WIDTH); // bottom-middle
@@ -124,22 +120,22 @@ void ::StoneyDSP::StoneyVCV::HP4Widget::draw(const ::rack::Widget::DrawArgs &arg
     // addChild(screwB1);
 }
 
-// ::StoneyDSP::StoneyVCV::HP4ModuleWidget::~HP4ModuleWidget()
+// ::StoneyDSP::StoneyVCV::HP1ModuleWidget::~HP1ModuleWidget()
 // {}
 
 /**
  *
  */
-::rack::plugin::Model* ::StoneyDSP::StoneyVCV::createHP4()
+::rack::plugin::Model* ::StoneyDSP::StoneyVCV::createHP1()
 {
 
-    ::rack::plugin::Model* modelHP4 = ::rack::createModel<
-        ::StoneyDSP::StoneyVCV::HP4Module,
-        ::StoneyDSP::StoneyVCV::HP4ModuleWidget
-    >("HP4");
+    ::rack::plugin::Model* modelHP1 = ::rack::createModel<
+        ::StoneyDSP::StoneyVCV::HP1Module,
+        ::StoneyDSP::StoneyVCV::HP1ModuleWidget
+    >("HP1");
 
-    // STONEYDSP_THROW_IF_FAILED_VOID(modelHP4 == nullptr, bad_alloc);
-    return modelHP4;
+    // STONEYDSP_THROW_IF_FAILED_VOID(modelHP1 == nullptr, bad_alloc);
+    return modelHP1;
 }
 
 //==============================================================================
@@ -148,7 +144,7 @@ namespace StoneyDSP {
 
 namespace StoneyVCV {
 
-::rack::plugin::Model* modelHP4 = ::StoneyDSP::StoneyVCV::createHP4();
+::rack::plugin::Model* modelHP1 = ::StoneyDSP::StoneyVCV::createHP1();
 
 }
